fix: Correction of MIN_MONOFORUM_CHANNEL_ID offset value
Corrected the definition of the MIN_MONOFORUM_CHANNEL_ID constant.
Previously, the offset value used + 1, which led to incorrect lower bound calculations.
This change replaces + 1 with - 1 to ensure MIN_MONOFORUM_CHANNEL_ID is calculated correctly,
according to the expected limits. This fixes a potential issue with channel ID validation or allocation.

Signed-off-by: Ling-ex <nekochan@rizkiofficial.com>